Separate CreateProjectDialogComponent into its own component
