+++ angular-app/src/app/create-project-dialog/create-project-dialog.component.html
@@ -1,0 +1,42 @@
+<mat-card>
+  <mat-card-header>
+    <mat-card-title>Create a New Project</mat-card-title>
+  </mat-card-header>
+  <mat-card-content>
+    <mat-form-field appearance="outline">
+      <mat-label>Project Title</mat-label>
+      <input matInput type="text" required [(ngModel)]="title">
+    </mat-form-field>
+    <br>
+    <mat-form-field>
+      <mat-label>Add Collaborators</mat-label>
+      <mat-chip-list #collaboratorList>
+        <mat-chip *ngFor="let collaborator of collaborators" 
+                      [removable]="true" (removed)="remove(collaborator)">
+          {{collaborator}}
+          <mat-icon matChipRemove>cancel</mat-icon>
+        </mat-chip>
+        <input 
+                    [matChipInputFor]="collaboratorList"
+                    (matChipInputTokenEnd)="addCollaborator($event)">
+      </mat-chip-list>
+    </mat-form-field>
+    <br>
+    <mat-form-field>
+      <mat-label>Add Tags</mat-label>
+      <mat-chip-list #tagList>
+        <mat-chip *ngFor="let tag of tags" 
+                    [removable]="true" (removed)="remove(tag)">
+          {{tag}}
+          <mat-icon matChipRemove>cancel</mat-icon>
+        </mat-chip>
+        <input 
+                    [matChipInputFor]="tagList"
+                    (matChipInputTokenEnd)="addTag($event)">
+      </mat-chip-list>
+    </mat-form-field>
+  </mat-card-content>
+  <mat-card-actions>
+    <button mat-button (click)='createProject()'>Create</button>
+  </mat-card-actions>
+</mat-card>
